requirements.txt
django
django-graphos
python-dateutil
